Initial commit of openstack-ops-messaging cookbook
This initial commit provides support for an openstack-ops-messaging
cookbook. In this initial release, we provide the capability to install
and configure RabbitMQ. This cookbook is modeled after the
openstack-ops-database cookbook in the way in which it can be utilized.

Included are tests for all recipes and this release supports Ubuntu
only.